+++ users.txt
@@ -1,1 +1,2 @@
 JavsRec05,javier,javi0204@gmail.com,1234
+blasfEma,Emanuel,ema@gmail.com,1234
